Bitmap and Path tests pass

+++ src/bitmap.cc
@@ -1,4 +1,6 @@
 #include "bitmap.hpp"
+#include <cstddef>
+#include <cassert>
 
 namespace ANAlloc {
 
@@ -17,19 +19,19 @@ Bitmap & Bitmap::operator=(const Bitmap & bm) {
   return *this;
 }
 
-Bitmap::Bitmap(uint8_t * _ptr, uintptr_t count) {
+Bitmap::Bitmap(uint8_t * _ptr, uint64_t count) {
   ptr = _ptr;
   bitCount = count;
-  for (uintptr_t i = 0; i < count / 8 + (count & 7 ? 1 : 0); i++) {
+  for (uint64_t i = 0; i < count / 8 + (count & 7 ? 1 : 0); i++) {
     ptr[i] = 0;
   }
 }
 
-void Bitmap::SetBit(uintptr_t idx, bool value) {
+void Bitmap::SetBit(uint64_t idx, bool value) {
   assert(idx < bitCount);
   
-  uintptr_t byteIndex = idx >> 3;
-  uintptr_t bitIndex = idx & 7;
+  uint64_t byteIndex = idx >> 3;
+  uint64_t bitIndex = idx & 7;
   if (!value) {
     ptr[byteIndex] &= 0xff ^ (1 << bitIndex);
   } else {
@@ -37,7 +39,7 @@ void Bitmap::SetBit(uintptr_t idx, bool value) {
   }
 }
 
-void Bitmap::SetMultibit(uintptr_t idx, int len, uintptr_t value) {
+void Bitmap::SetMultibit(uint64_t idx, int len, uint64_t value) {
   // TODO: optimize this
   assert(idx + len <= bitCount);
   for (int i = 0; i < len; i++) {
@@ -46,26 +48,26 @@ void Bitmap::SetMultibit(uintptr_t idx, int len, uintptr_t value) {
   }
 }
 
-bool Bitmap::GetBit(uintptr_t idx) const {
+bool Bitmap::GetBit(uint64_t idx) const {
   assert(idx < bitCount);
   
-  uintptr_t byteIndex = idx >> 3;
-  uintptr_t bitIndex = idx & 7;
+  uint64_t byteIndex = idx >> 3;
+  uint64_t bitIndex = idx & 7;
   return (ptr[byteIndex] & (1 << bitIndex)) != 0;
 }
 
-uintptr_t Bitmap::GetMultibit(uintptr_t idx, int len) const {
+uint64_t Bitmap::GetMultibit(uint64_t idx, int len) const {
   // there are len bits, and the one at idx is the most significant
   // TODO: figure out if this can be optimized
   assert(idx + len <= bitCount);
-  uintptr_t result = 0;
+  uint64_t result = 0;
   for (int i = 0; i < len; i++) {
-    result += (uintptr_t)GetBit(i + idx) << (len - i - 1);
+    result += (uint64_t)GetBit(i + idx) << (len - i - 1);
   }
   return result;
 }
 
-uintptr_t Bitmap::GetBitCount() const {
+uint64_t Bitmap::GetBitCount() const {
   return bitCount;
 }
